% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{examine_dataset}
\alias{examine_dataset}
\title{Generate several plots and objects for dataset exploration}
\usage{
examine_dataset(cell_data_dataframe = data.frame(),
  normalized_expression_matrix = matrix(), .species = "mmusculus",
  color_var = character())
}
\arguments{
\item{cell_dataframe:}{cell_dataframe consists of sample names of each cell
to be examined, as well as all other data that might be used for
visualizing the resulting cells. This data.frame must have a character
column named 'sample_name', which must match the row names of
normalized_expression_matrix}

\item{normalized_expression_matrix:}{this must be a CxG matrix, with C cells
and G genes, The C cells are the rownamews of the matrix and must match the
sample_name column from cell_dataframe. The column names must be ENSEMBL
IDs. The entries of the matrix must be the normalized since they will be
used in tsne}

\item{.species:}{currently either 'mmusculus' or 'hsapiens'}

\item{color_var:}{variable name to color cells by in overview ggplot
images. This must correspond to a column name in cell_data_dataframe.}
}
\value{
This function returns a list with the following elements: \itemize{
  \item \strong{tsne_results}: a list of Rtsne outputs form the filtered
  dataset, each from a differnet perplexity (15,30,45,60) \item
  \strong{n50_enriched_genes_by_cluster}: a dataframe showing the top 50
  enriched genes in each cluster. Enrichment is based on proportion of cells
  expressing a gene \item \strong{clusters_scatterplot}: a scatterplot coloring the cells by cluster_id\item
  \strong{important_genes_heatmap}: a heatmap plot showing the enriched genes in each cluster \item \strong{overview_plots}: a list of plots corresponding to the tsne runs that are coloredc by @param color_by
}
}
\description{
This function examines several facets of the dataset, inclusing filtering
cells based on metadata, dimension reduction, cluster analysis, and
visualization
}
