% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{explore_expression}
\alias{explore_expression}
\title{View expression of genes on 2D plots.}
\usage{
explore_expression(ggplot_df, expr_mat, gene_selector_df)
}
\arguments{
\item{ggplot_df}{a data.frame that must contain the following column names and associated types: sample_name, character; D1, numeric; D2, numeric}

\item{expr_mat}{a matrix with rownames corresponding to sample_name and with column names corresponding to ENSEMBL IDs. The matrix must be numeric}

\item{gene_selector_df}{a data.frame with an ENSEMBL column corresponding to the colnames of expr_mat and containing other informative columns such as GENENAME, GENE_TYPE, etc...}
}
\value{
Returns null.
}
\description{
View expression of genes on 2D plots.
}
