% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_cellranger_matrices}
\alias{get_cellranger_matrices}
\title{Loads the matrix files from a 10X run}
\usage{
get_cellranger_matrices(filenames = character(), sample_name_prefix = NULL,
  genome_build = "hg19")
}
\value{
get_cellranger_matrices returns a list of count matrices corresponding to the matrix files of \code{filenames}. The names of the list are the basenames of \code{filenames}
}
\description{
Loads the matrix files from a 10X run
}

