% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_cidr_distance_matrix_from_raw_counts}
\alias{get_cidr_distance_matrix_from_raw_counts}
\title{Compute distance between samples}
\usage{
get_cidr_distance_matrix_from_raw_counts(cell_by_ensembl_mat = matrix())
}
\value{
get_cidr_distance_matrix_from_raw_counts returns a named distance matrix showing the distance between samples. Both rows and columns are named.
}
\description{
Use cidr to compute inter-sample distances and return named distance matrix
}
