% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{get_cluster_by_ensembl_mat_of_expr_summaries}
\alias{get_cluster_by_ensembl_mat_of_expr_summaries}
\title{calculate cluster summaries}
\usage{
get_cluster_by_ensembl_mat_of_expr_summaries(sample_by_ensembl_mat,
  sample_to_cluster_df = data.frame(), summary_function)
}
\value{
get_cluster_by_ensembl_mat_of_expr_summaries retuens a cluster by ensembl matrix consisting of the summmary value for each cluster and ensembl combination.
}
\description{
calculate summary for each cluster defined in sample_to_cluster_df and output summary results.
}

