% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{get_mean_dist_from_matrix}
\alias{get_mean_dist_from_matrix}
\title{The function calculates and returns the mean of the upper-triangular portion of a matrix}
\usage{
get_mean_dist_from_matrix(temp_mat)
}
\value{
get_mean_dist_from_matrix returns the upper-triangular mean of temp_mat, exluding the zeros from the diagonal or the lower-triangular portion of the matrix
}
\description{
The function calculates and returns the mean of the upper-triangular portion of a matrix
}

