% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{get_random_genes_from_matrix}
\alias{get_random_genes_from_matrix}
\title{Filter random genes from matrix}
\usage{
get_random_genes_from_matrix(count_matrix = matrix(),
  recompute_dist_removal_threshold = 2000, min_cell_count_threshold = 4,
  distance_matrix = NULL, distance_function = dist,
  normalize_matrix = TRUE, heuristic_threshold = 200,
  heuristic_step_size = 20, n_sample_draws = 1000, tail_area = 0.2)
}
\description{
Remove genes from raw count matrix that are unlikely contributing to the biology based on their expression patterns
}
\details{
This function performs the following filtering tasks. It initially removes all columns (genes) that have less than two non-zero entries, ie. that are expressed in zero or one cell.
This c
}

