% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{read_merge_and_return_sample_matrices}
\alias{read_merge_and_return_sample_matrices}
\title{Loads and merges the matrix files from a 10X run}
\usage{
read_merge_and_return_sample_matrices(filenames = character(),
  sample_name_prefix = NULL, genome_build = "hg19")
}
\value{
read_merge_and_return_sample_matrices returns a count matrices corresponding to the matrix files of \code{filenames}. The matrices of the original samples are merged with the rows corresponding to samples and the columns corresponding to ensembl ids.
}
\description{
Loads and merges the matrix files from a 10X run
}

