% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{run_interactive_monocle_gadget}
\alias{run_interactive_monocle_gadget}
\title{Explore a dataset with monocle}
\usage{
run_interactive_monocle_gadget(expression_mat = matrix(),
  max_cell_counts = 1000, metadata_df = data.frame(),
  species = c("mmusculus", "hsapiens"), normalizing_scale_factor = NULL,
  cell_umi_counts = NULL)
}
\arguments{
\item{expression_mat:}{This is a CellxGene matrix that will be used to run
monocle.}

\item{max_cell_counts:}{Moncole can choke on large datasets, so this puts a
cap on then number of cells from expression_mat that will be used to
generate a 2D structure.}

\item{metadata_df:}{This is a data.frame that contains metatadata for the
cells in expression_mats. It is used for coloring the monocle plots by
additinal variables. It must at the very least have a sample_name column
that matches the rownames of expression_mat.}

\item{species:}{this should be pretty obvious. Currently, 'mmusculus' and
'hsapiens' are valid}

\item{normalizing_scale_factor:}{This is to un-normalize the data. This has
to be done because of the different normalization procedure that monocle
uses. This factor is just the median umi count that was used in the
normalization of expression_mat. It can be found as an element of the list
returned from get_normalized_expression_matrix.}

\item{cell_umi_counts:}{This is a named character vector of cell UMI counts,
again required for un-normalizing the data in expression_mat. It can also
be found as an element of the list returned from
get_normalized_expression_matrix.}
}
\value{
This function launches an app that allows interactive exploration of
  a dataset. There is no object returned.
}
\description{
This gadget uses monocle to plot a DDRTree and visualization of a branched
heatmap. The branch point and root of the tree can be changed interactively.
}
